Implemented real-time preview and real-time validation. Close #33 and close #12.
Now input commands will be parsed and validatted as they are typed, and the mistakes will be displayed just below the timer area.
Mistakes in the last command will not be displayed because the command is not finished, and the detected error may be corrected as the user type.
And the content of the first command is previewed as soon as the least information is given, that is the command name (down or up) and the duration such as 05:00.