Fixed the parser for commands. Close #40.
The issue was caused by the `pos` parser, which fails when the position is at the end of the input string. In this case, `pos` was used with the `end'` parser, which succeeds when the position is as the end of the input string. Those two can't be used tohether.
Now a new parser `pos'` is defined. This version doesn't fail even when the parsing position comes to the end.